requirements.txt
Click==7.0
Flask==1.0.2
flask-msearch==0.1.9
Flask-SQLAlchemy==2.3.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
Werkzeug==0.14.1

